+++ src/config/configServer.cpp
@@ -6,7 +6,7 @@
 /*   By: fkoolhov <fkoolhov@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/18 18:27:26 by felicia           #+#    #+#             */
-/*   Updated: 2024/04/25 12:53:37 by fkoolhov         ###   ########.fr       */
+/*   Updated: 2024/04/25 14:39:05 by fkoolhov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,7 +55,7 @@ static void get_custom_error_page_from_config(std::unique_ptr<Server>& server, s
 		i--;
 		std::string error_page = words[i];
 		for (unsigned long j = 1; j < i; j++)
-			server->addCustomErrorPage(std::stoi(words[j]), error_page);
+			server->addCustomErrorPage(static_cast<short>(std::stoi(words[j])), error_page);
 	}
 }
 
@@ -171,7 +171,7 @@ static void create_full_server_paths(std::unique_ptr<Server>& server)
 {
 	server->setRootFolder("./" + server->getRootFolder());
 	server->setDefaultErrorPage(server->getRootFolder() + server->getDefaultErrorPage());
-	std::map<int, std::string> custom_error_pages = server->getCustomErrorPages();
+	std::map<short, std::string> custom_error_pages = server->getCustomErrorPages();
 	for (auto it = custom_error_pages.begin(); it != custom_error_pages.end(); it++)
 		server->addCustomErrorPage(it->first, server->getRootFolder() + it->second);
 }
